# openapi/erp-api.yaml
openapi: 3.1.0
info:
  title: ERP Platform API
  description: >-
    Modular monolith ERP. Operations marked x-implementation-status: roadmap are contracts only and are NOT implemented.
    Public routes: health, status, register and login. Refresh/logout require a refresh credential. Login requires x-tenant-id. Protected routes require JWT plus
    matching tenant context and an active membership. No default tenant. Responses use data/meta/correlationId or
    error/correlationId, without a success flag. Company/branch headers are contextual metadata, not independently
    authorized scopes. Business company routes also require an active persisted company and explicit company membership.
  version: v1
  contact:
    name: ERP Platform Team
    email: dev@erp.local
    url: https://erp.local
  license:
    name: Proprietary
    url: https://erp.local/license
servers:
  - url: http://localhost:3000
    description: Local development; paths include /api/v1 where applicable
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        HS256 access JWT with sub, sid, tenantId, roles, iat and exp. Session, user and membership are checked in
        MongoDB on every request. Revocation denies subsequent checks immediately; in-flight work is not rolled back.
    refreshCookie:
      type: apiKey
      in: cookie
      name: erp_refresh
      description: HttpOnly SameSite Strict cookie scoped to /api/v1/auth; Secure in production. Web writes require an exact trusted Origin and x-session-client web.
  schemas:
    Error:
      type: object
      required:
        - error
        - correlationId
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        correlationId:
          type: string
      example:
        error:
          code: FORBIDDEN
          message: Forbidden
        correlationId: example-request-id
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_SERVER_ERROR
            - BAD_REQUEST
            - TENANT_ACCESS_DENIED
            - INSUFFICIENT_STOCK
            - INVENTORY_INSUFFICIENT_STOCK
            - UNAUTHORIZED_TRANSITION
            - FISCAL_PERIOD_CLOSED
            - DUPLICATE_RESOURCE
            - RESOURCE_NOT_FOUND
            - VALIDATION_SCHEMA_ERROR
            - IDEMPOTENCY_KEY_EXISTS
            - RATE_LIMIT_EXCEEDED
            - EXTERNAL_SERVICE_ERROR
            - TIMEOUT_ERROR
            - FEATURE_DISABLED
            - PERMISSION_DENIED
            - TENANT_CONTEXT_MISSING
            - INTERNAL_ERROR
            - RATE_LIMITED
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
...
    ProductCreate:
      type: "object"
      additionalProperties: false
      required: ["name","sku","unitOfMeasure"]
      properties:
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        sku:
          type: "string"
          maxLength: 200
          minLength: 1
          description: "Normalized uppercase; unique within tenant and company"
        unitOfMeasure:
          type: "string"
          enum: ["piece","kg","liter","meter","box","set"]
        description:
          type: "string"
          maxLength: 2000
        category:
          type: "string"
          maxLength: 2000
        subcategory:
          type: ["string","null"]
          maxLength: 2000
        costPrice:
          type: "number"
          minimum: 0
        sellingPrice:
          type: "number"
          minimum: 0
        taxRate:
          type: "number"
          minimum: 0
          description: "Reference metadata only; not a fiscal tax calculation"
        weight:
          type: ["number","null"]
          minimum: 0
        dimensions:
          oneOf: [{"type":"null"},{"type":"object","additionalProperties":false,"required":["length","width","height","unit"],"properties":{"length":{"type":"number","minimum":0},"width":{"type":"number","minimum":0},"height":{"type":"number","minimum":0},"unit":{"type":"string","maxLength":20}}}]
        isActive:
          type: "boolean"
    ProductPatch:
      type: "object"
      additionalProperties: false
      required: ["expectedVersion"]
      properties:
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        sku:
          type: "string"
          maxLength: 200
          minLength: 1
          description: "Normalized uppercase; unique within tenant and company"
        unitOfMeasure:
          type: "string"
          enum: ["piece","kg","liter","meter","box","set"]
        description:
          type: "string"
          maxLength: 2000
        category:
          type: "string"
          maxLength: 2000
        subcategory:
          type: ["string","null"]
          maxLength: 2000
        costPrice:
          type: "number"
          minimum: 0
        sellingPrice:
          type: "number"
          minimum: 0
        taxRate:
          type: "number"
          minimum: 0
          description: "Reference metadata only; not a fiscal tax calculation"
        weight:
          type: ["number","null"]
          minimum: 0
        dimensions:
          oneOf: [{"type":"null"},{"type":"object","additionalProperties":false,"required":["length","width","height","unit"],"properties":{"length":{"type":"number","minimum":0},"width":{"type":"number","minimum":0},"height":{"type":"number","minimum":0},"unit":{"type":"string","maxLength":20}}}]
        isActive:
          type: "boolean"
        expectedVersion:
          type: "integer"
          minimum: 0
          maximum: 9007199254740990
    Product:
      type: "object"
      additionalProperties: false
      required: ["id","companyId","name","sku","unitOfMeasure","isActive","version","createdAt","updatedAt"]
      properties:
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        sku:
          type: "string"
          maxLength: 200
          minLength: 1
          description: "Normalized uppercase; unique within tenant and company"
        unitOfMeasure:
          type: "string"
          enum: ["piece","kg","liter","meter","box","set"]
        description:
          type: "string"
          maxLength: 2000
        category:
          type: "string"
          maxLength: 2000
        subcategory:
          type: ["string","null"]
          maxLength: 2000
        costPrice:
          type: "number"
          minimum: 0
        sellingPrice:
          type: "number"
          minimum: 0
        taxRate:
          type: "number"
          minimum: 0
          description: "Reference metadata only; not a fiscal tax calculation"
        weight:
          type: ["number","null"]
          minimum: 0
        dimensions:
          oneOf: [{"type":"null"},{"type":"object","additionalProperties":false,"required":["length","width","height","unit"],"properties":{"length":{"type":"number","minimum":0},"width":{"type":"number","minimum":0},"height":{"type":"number","minimum":0},"unit":{"type":"string","maxLength":20}}}]
        isActive:
          type: "boolean"
        id:
          type: "string"
          format: "uuid"
        companyId:
          type: "string"
          format: "uuid"
        version:
          type: "integer"
          minimum: 0
          maximum: 9007199254740990
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
    WarehouseCreate:
      type: "object"
      additionalProperties: false
      required: ["name","code"]
      properties:
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        code:
          type: "string"
          maxLength: 200
          minLength: 1
          description: "Uppercase; unique within tenant/company"
        location:
          type: "string"
          maxLength: 1000
        type:
          type: "string"
          enum: ["main","regional","local","dropship"]
        capacity:
          type: "number"
          minimum: 0
        isActive:
          type: "boolean"
    WarehousePatch:
      type: "object"
      additionalProperties: false
      required: ["expectedVersion"]
      properties:
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        code:
          type: "string"
          maxLength: 200
          minLength: 1
          description: "Uppercase; unique within tenant/company"
        location:
          type: "string"
          maxLength: 1000
        type:
          type: "string"
          enum: ["main","regional","local","dropship"]
        capacity:
          type: "number"
          minimum: 0
        isActive:
          type: "boolean"
        expectedVersion:
          type: "integer"
          minimum: 0
          maximum: 9007199254740990
    Warehouse:
      type: "object"
      additionalProperties: false
      required: ["id","companyId","name","code","isActive","version","createdAt","updatedAt"]
      properties:
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        code:
          type: "string"
          maxLength: 200
          minLength: 1
          description: "Uppercase; unique within tenant/company"
        location:
          type: "string"
          maxLength: 1000
        type:
          type: "string"
          enum: ["main","regional","local","dropship"]
        capacity:
          type: "number"
          minimum: 0
        isActive:
          type: "boolean"
        id:
          type: "string"
          format: "uuid"
        companyId:
          type: "string"
          format: "uuid"
        version:
          type: "integer"
          minimum: 0
          maximum: 9007199254740990
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
...
    ProductPage:
      type: "object"
      additionalProperties: false
      required: ["items","total","page","limit","totalPages"]
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: "integer"
          minimum: 0
        page:
          type: "integer"
          minimum: 1
        limit:
          type: "integer"
          minimum: 1
          maximum: 100
        totalPages:
          type: "integer"
          minimum: 0
    WarehousePage:
      type: "object"
      additionalProperties: false
      required: ["items","total","page","limit","totalPages"]
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/Warehouse"
        total:
          type: "integer"
          minimum: 0
        page:
          type: "integer"
          minimum: 1
        limit:
          type: "integer"
          minimum: 1
          maximum: 100
        totalPages:
          type: "integer"
          minimum: 0
...
  parameters:
...
    CompanyIdPathParam:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The company ID
...
    TenantIdHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
      description: >-
        Explicit tenant selector. Must equal the JWT tenant on protected routes; missing returns 401
        TENANT_CONTEXT_MISSING, mismatch returns 403.
...
paths:
...
  "/api/v1/companies/{companyId}/inventory/products":
    get:
      operationId: "listProducts"
      summary: "List products"
      description: "Requires inventory.product.read. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. Stable createdAt/id ordering."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"},{"name":"page","in":"query","schema":{"type":"integer","minimum":1,"maximum":1000000,"default":1}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":100,"default":50}},{"name":"q","in":"query","schema":{"type":"string","maxLength":200}}]
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/ProductPage"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: "createProduct"
      summary: "Create products"
      description: "Requires inventory.product.write. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. Business mutation and audit are transactional."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"}]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/Product"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/companies/{companyId}/inventory/products/{id}":
    get:
      operationId: "getProduct"
      summary: "Read product"
      description: "Requires inventory.product.read. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. "
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/Product"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: "updateProduct"
      summary: "Edit product"
      description: "Requires inventory.product.write. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. At least one mutable field plus expectedVersion. No stock changes through catalog edits."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/Product"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/companies/{companyId}/inventory/warehouses":
    get:
      operationId: "listWarehouses"
      summary: "List warehouses"
      description: "Requires inventory.warehouse.read. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. Stable createdAt/id ordering."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"},{"name":"page","in":"query","schema":{"type":"integer","minimum":1,"maximum":1000000,"default":1}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":100,"default":50}},{"name":"q","in":"query","schema":{"type":"string","maxLength":200}}]
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/WarehousePage"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: "createWarehouse"
      summary: "Create warehouses"
      description: "Requires inventory.warehouse.write. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. Business mutation and audit are transactional."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"}]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/WarehouseCreate"
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/Warehouse"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/companies/{companyId}/inventory/warehouses/{id}":
    get:
      operationId: "getWarehouse"
      summary: "Read warehouse"
      description: "Requires inventory.warehouse.read. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. "
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/Warehouse"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: "updateWarehouse"
      summary: "Edit warehouse"
      description: "Requires inventory.warehouse.write. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. At least one mutable field plus expectedVersion. No stock changes through catalog edits."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/WarehousePatch"
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/Warehouse"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
...